update popup trailer
