--- src/pages/Homepage/Carousel/popup/css.css
@@ -1,4 +1,0 @@
-.MuiDialog-paper {
-  overflow-y: visible !important;
-  background-color: black !important;
-}
